// src/nft-metadata-generator/nft-metadata-generator.enums.ts
export enum Tier {
  Miner = 'Miner',
}

export enum Background {
  Mystical = 'Mystical',
  Dark = 'Dark',
  Misty = 'Misty',
  Jade = 'Jade',
  Light = 'Light',
}

export enum Eyes {
  Glowing = 'Glowing',
  Focused = 'Focused',
  Tired = 'Tired',
  Closed = 'Closed',
  Serious = 'Serious',
}

export enum Mouth {
  JadeTooth = 'Jade Tooth',
  Laughing = 'Laughing',
  Whistling = 'Whistling',
  Smirking = 'Smirking',
  Smiling = 'Smiling',
}

export enum Hair {
  JadeStreaks = 'Jade Streaks',
  Shaved = 'Shaved',
  Curly = 'Curly',
  Braided = 'Braided',
  Short = 'Short',
}

export enum Clothes {
  JadeVeinVest = 'Jade Vein Vest',
  MinersTank = `Miner's Tank`,
  GemstoneGambeson = 'Gemstone Gambeson',
  DustCoat = 'Dust Coat',
  Engineersjacket = 'Engineers jacket',
}